Separate all test code (except for 'DummyRtlSdr') outside of the library code
