+++ pytest.ini
@@ -1,3 +1,3 @@
 [pytest]
-testpaths = rtlsdr
+testpaths = tests
 python_files = *.py
