+++ tests/test_basic.py
@@ -1,0 +1,6 @@
+
+def test(sdr_cls):
+    from utils import generic_test
+    sdr = sdr_cls()
+    generic_test(sdr)
+    sdr.close()
